Refactoring using spring and exposing APIs
