+++ src/main/java/com/smarthome/HomeApplianceSystemApplication.java
@@ -1,0 +1,11 @@
+package com.smarthome;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HomeApplianceSystemApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(HomeApplianceSystemApplication.class, args);
+    }
+}
