+++ src/main/java/com/smarthome/appliance/Appliance.java
@@ -1,5 +1,8 @@
-package appliances;
+package com.smarthome.appliance;
 
+import org.springframework.stereotype.Component;
+
+@Component
 public interface Appliance {
 
     void turnOn();
