+++ src/main/java/com/smarthome/commands/Command.java
@@ -1,4 +1,4 @@
-package commands;
+package com.smarthome.commands;
 
 public interface Command {
