+++ src/main/java/com/smarthome/enums/Switch.java
@@ -1,4 +1,4 @@
-package enums;
+package com.smarthome.enums;
 
 public enum Switch {
     ON, OFF
