+++ src/test/java/SystemUpdateTest.java
@@ -1,35 +1,18 @@
-import controller.SmartHomeApplianceController;
+import com.smarthome.controller.ApplianceRegistry;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import scheduler.HomeApplianceSystemUpdate;
-
-import java.util.concurrent.*;
 
 import static org.mockito.Mockito.*;
 
-class SystemUpdateTest {
+class HomeApplianceSystemUpdateTest {
 
     @Test
-    void testUpdateTaskExecutesAndTurnsOffAllAppliances() throws InterruptedException {
-        SmartHomeApplianceController controller = mock(SmartHomeApplianceController.class);
+    void testSystemUpdateTurnsOffAllAppliances() {
+        ApplianceRegistry controller = mock(ApplianceRegistry.class);
+        HomeApplianceSystemUpdate updateSystem = new HomeApplianceSystemUpdate(controller);
 
-        ScheduledExecutorService testScheduler = Executors.newSingleThreadScheduledExecutor();
-        HomeApplianceSystemUpdate updateSystem = new HomeApplianceSystemUpdate(controller, testScheduler);
+        updateSystem.runSystemUpdateLogic();
 
-        // Replace scheduleNextUpdate with a short one-time delay
-        Runnable testTask = () -> {
-            System.out.println("Simulating system update...");
-            controller.turnOffAll();
-        };
-
-        // Act: Schedule task to run in 100 ms
-        testScheduler.schedule(testTask, 100, TimeUnit.MILLISECONDS);
-
-        // Wait briefly to ensure task runs
-        Thread.sleep(200);
-
-        // Assert
         verify(controller, times(1)).turnOffAll();
-        testScheduler.shutdownNow();
     }
 }
